Kung fu po cambios diseño usuario (#639)
* Feat:Estilos de celda agregado

* Feat:Refact con los estilos de celda

* refactor: extraer header y tabla en componentes reutilizables en usuarios

* refactor(users): extraer lógica de usuarios a custom hook
