+++ Web Ui/src/sections/User/UserPage.tsx
@@ -1,47 +1,13 @@
-import { useState, useEffect, useMemo } from "react";
-
-import GetUsers from "../../modules/Users/application/getUsers";
-import UsersRepository from "../../modules/Users/repository/UsersRepository";
-import { UserDataObject } from "../../modules/Users/domain/UsersInterface";
-import { RemoveUserFromGroup } from "../../modules/Users/application/removeUserFromGroup";
-
-import {
-  Table,TableHead,TableBody,TableRow,TableCell,Container,Select,MenuItem,FormControl,CircularProgress,SelectChangeEvent,Tooltip,TableContainer,TextField,InputAdornment,
-} from "@mui/material";
-
+import { Container, CircularProgress, TableContainer } from "@mui/material";
 import { styled } from "@mui/system";
-import RemoveCircleIcon from "@mui/icons-material/RemoveCircle";
-import SearchIcon from "@mui/icons-material/Search";
 
-import GetGroups from "../../modules/Groups/application/GetGroups";
-import { GroupDataObject } from "../../modules/Groups/domain/GroupInterface";
-import GroupsRepository from "../../modules/Groups/repository/GroupsRepository";
-
-import { SearchUsersByEmail } from "../../modules/Users/application/SearchUsersByEmail";
+import UsersHeader from "./UsersHeader";
+import UsersTable from "./UsersTable";
+import { useUsersPageData } from "./useUsersPageData";
 
 // ------------------- ESTILOS -------------------
 const CenteredContainer = styled(Container)({
   marginTop: "28px",
-});
-
-const HeaderContainer = styled("div")({
-  width: "82%",
-  margin: "0 auto",
-  display: "flex",
-  justifyContent: "space-between",
-  alignItems: "center",
-});
-
-const HeaderFilters = styled("div")({
-  display: "flex",
-  alignItems: "center",
-  gap: "16px",
-});
-
-const Title = styled("h2")({
-  fontSize: "18px",
-  fontWeight: 600,
-  margin: 0,
 });
 
 const DividerLine = styled("div")({
@@ -58,102 +24,21 @@ const StyledTableContainer = styled(TableContainer)({
   boxShadow: "none",
   backgroundColor: "#fff",
 });
-
-const StyledTable = styled(Table)({
-  width: "100%",
-  borderCollapse: "collapse",
-});
 // -------------------------------------------------
 
 function UserPage() {
-  const [, setUsers] = useState<UserDataObject[]>([]);
-  const [groups, setGroups] = useState<GroupDataObject[]>([]);
-  const [selectedGroup, setSelectedGroup] = useState<number | "all">("all");
-  const [searchQuery, setSearchQuery] = useState("");
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState<unknown>(null);
-  const [filteredUsers, setFilteredUsers] = useState<UserDataObject[]>([]);
+  const {
+    groups, selectedGroup, searchQuery, loading, error,
+    filteredUsers, groupMap, setSearchQuery, handleGroupChange, handleRemoveUserFromGroup,
+  } = useUsersPageData();
 
-  // --- INSTANCIAS ---
-  const userRepository = useMemo(() => new UsersRepository(), []);
-  const getUsers = useMemo(() => new GetUsers(userRepository), [userRepository]);
-  const getGroups = useMemo(() => new GetGroups(new GroupsRepository()), []);
-  const searchUsersByEmail = useMemo(
-    () => new SearchUsersByEmail(userRepository),
-    [userRepository]
-  );
-
-  // ------------------- FETCH USERS + GROUPS -------------------
-  useEffect(() => {
-    const fetchUsersAndGroups = async () => {
-      try {
-        const [userData, groupData] = await Promise.all([
-          getUsers.getUsers(),
-          getGroups.getGroups(),
-        ]);
-
-        setUsers(userData);
-        setGroups(groupData);
-      } catch (error) {
-        console.error("Error fetching users or groups:", error);
-        setError(error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchUsersAndGroups();
-  }, [getUsers, getGroups]);
-
-  // ------------------- FILTRO DE USUARIOS -------------------
-  useEffect(() => {
-    const runSearch = async () => {
-      const results = await searchUsersByEmail.execute({
-        query: searchQuery,
-        groupId: selectedGroup,
-      });
-
-      setFilteredUsers(results);
-    };
-
-    runSearch();
-  }, [searchQuery, selectedGroup, searchUsersByEmail]);
-
-  // ------------------- MAPA DE GRUPOS -------------------
-  const groupMap = groups.reduce((acc, group) => {
-    acc[group.id] = group.groupName;
-    return acc;
-  }, {} as { [key: number]: string });
-
-  // ------------------- HANDLERS -------------------
-  const handleGroupChange = (event: SelectChangeEvent<number | "all">) => {
-    setSelectedGroup(event.target.value as number | "all");
-  };
-
-  const handleRemoveUserFromGroup = async (userId: number) => {
-    if (window.confirm("¿Estás seguro que deseas eliminar del grupo a este estudiante?")) {
-      try {
-        const removeUserInstance = new RemoveUserFromGroup(userRepository);
-        await removeUserInstance.removeUserFromGroup(userId);
-        alert("Estudiante eliminado con éxito del grupo.");
-        window.location.reload();
-      } catch (error) {
-        console.error(error);
-        alert("Hubo un error al eliminar al estudiante del grupo.");
-      }
-    }
-  };
-
-  // ------------------- RENDER -------------------
   if (loading) {
     return (
       <div
         style={{
           display: "flex",
           justifyContent: "center",
-          alignItems: "center",
           height: "100vh",
-          width: "100vw",
         }}
       >
         <CircularProgress />
@@ -161,165 +46,30 @@ function UserPage() {
     );
   }
 
-  if (error) return <div>Error: {(error as Error).message}</div>;
+  if (error) {
+    return <div>Error: {(error as Error).message}</div>;
+  }
 
   return (
-    <div>
-      <CenteredContainer>
-        <HeaderContainer>
-          <Title>Usuarios</Title>
+    <CenteredContainer>
+      <UsersHeader
+        searchQuery={searchQuery}
+        selectedGroup={selectedGroup}
+        groups={groups}
+        onSearchChange={setSearchQuery}
+        onGroupChange={handleGroupChange}
+      />
 
-          <HeaderFilters>
-            <TextField
-              placeholder="Buscar por email"
-              size="small"
-              value={searchQuery}
-              onChange={(e) => setSearchQuery(e.target.value)}
-              sx={{ width: 260 }}
-              InputProps={{
-                startAdornment: (
-                  <InputAdornment position="start">
-                    <SearchIcon />
-                  </InputAdornment>
-                ),
-              }}
-            />
+      <DividerLine />
 
-            <FormControl size="small" sx={{ minWidth: 180 }}>
-              <Select
-                value={selectedGroup}
-                onChange={handleGroupChange}
-                displayEmpty
-              >
-                <MenuItem value="all">Filtrar todos los grupos</MenuItem>
-                {groups.map((group) => (
-                  <MenuItem key={group.id} value={group.id}>
-                    {group.groupName}
-                  </MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-          </HeaderFilters>
-        </HeaderContainer>
-
-        <DividerLine />
-
-        <section className="Usuarios" style={{ width: "100%" }}>
-          <StyledTableContainer>
-            <StyledTable>
-              <TableHead>
-                <TableRow>
-                  <TableCell
-                    sx={{
-                      fontWeight: 600,
-                      borderBottom: "1px solid #d9d9d9",
-                      borderRight: "1px solid #d9d9d9",
-                      padding: "10px 8px",
-                    }}
-                  >
-                    Correo
-                  </TableCell>
-                  <TableCell
-                    sx={{
-                      fontWeight: 600,
-                      borderBottom: "1px solid #d9d9d9",
-                      borderRight: "1px solid #d9d9d9",
-                      padding: "10px 8px",
-                    }}
-                  >
-                    Grupo
-                  </TableCell>
-                  <TableCell
-                    sx={{
-                      fontWeight: 600,
-                      borderBottom: "1px solid #d9d9d9",
-                      borderRight: "1px solid #d9d9d9",
-                      padding: "10px 8px",
-                    }}
-                  >
-                    Rol
-                  </TableCell>
-                  <TableCell
-                    sx={{
-                      fontWeight: 600,
-                      borderBottom: "1px solid #d9d9d9",
-                      padding: "10px 8px",
-                    }}
-                  >
-                    Eliminar
-                  </TableCell>
-                </TableRow>
-              </TableHead>
-
-              <TableBody>
-                {filteredUsers.length === 0 ? (
-                  <TableRow>
-                    <TableCell
-                      colSpan={4}
-                      sx={{
-                        textAlign: "center",
-                        py: 3,
-                        borderBottom: "1px solid #d9d9d9",
-                      }}
-                    >
-                      No se encontraron resultados
-                    </TableCell>
-                  </TableRow>
-                ) : (
-                  filteredUsers.map((user) => (
-                    <TableRow key={user.id}>
-                      <TableCell
-                        sx={{
-                          borderBottom: "1px solid #d9d9d9",
-                          borderRight: "1px solid #d9d9d9",
-                          padding: "10px 8px",
-                        }}
-                      >
-                        {user.email}
-                      </TableCell>
-                      <TableCell
-                        sx={{
-                          borderBottom: "1px solid #d9d9d9",
-                          borderRight: "1px solid #d9d9d9",
-                          padding: "10px 8px",
-                        }}
-                      >
-                        {groupMap[user.groupid] || "Unknown"}
-                      </TableCell>
-                      <TableCell
-                        sx={{
-                          borderBottom: "1px solid #d9d9d9",
-                          borderRight: "1px solid #d9d9d9",
-                          padding: "10px 8px",
-                        }}
-                      >
-                        {user.role}
-                      </TableCell>
-                      <TableCell
-                        sx={{
-                          borderBottom: "1px solid #d9d9d9",
-                          padding: "10px 8px",
-                        }}
-                      >
-                        <Tooltip title={`Eliminar de ${groupMap[user.groupid]}`} arrow>
-                          <RemoveCircleIcon
-                            onClick={() => handleRemoveUserFromGroup(user.id)}
-                            sx={{
-                              color: "#ff1a1a",
-                              cursor: "pointer",
-                            }}
-                          />
-                        </Tooltip>
-                      </TableCell>
-                    </TableRow>
-                  ))
-                )}
-              </TableBody>
-            </StyledTable>
-          </StyledTableContainer>
-        </section>
-      </CenteredContainer>
-    </div>
+      <StyledTableContainer>
+        <UsersTable
+          users={filteredUsers}
+          groupMap={groupMap}
+          onRemoveUser={handleRemoveUserFromGroup}
+        />
+      </StyledTableContainer>
+    </CenteredContainer>
   );
 }
